successfully created user with register page
