+++ src/actions/auth.actions.ts
@@ -2,8 +2,6 @@
 
 import { signIn } from "@/lib/auth";
 import { AuthError } from "next-auth";
-import { dbConnect } from "@/lib/db";
-import { User } from "@/models/User.model";
 
 export async function authenticate(
   prevState: string | undefined,
@@ -22,34 +20,4 @@ export async function authenticate(
     }
     throw error;
   }
-}
-
-// register user function
-export async function registerUser(
-  prevState: string | undefined,
-  formData: FormData
-) {
-  console.log(formData);
-  await dbConnect();
-
-  const firstName = formData.get("firstName") as string;
-  const lastName = formData.get("lastName") as string;
-  const email = formData.get("email") as string;
-  const password = formData.get("password") as string;
-
-  if (!firstName || !lastName || !email || !password) {
-    return "Please fill in all fields";
-  }
-
-  const newUser = new User({
-    firstName,
-    lastName,
-    email,
-    password,
-    role: "user",
-  });
-
-  await newUser.save();
-
-  return "User created successfully";
 }
